task-2/requirements.txt
# requirements.txt
flask==2.0.1
redis
paho-mqtt
pyjwt
Flask-PyMongo
Werkzeug==2.2.2
